Renaming resolvers
