+++ lib/plugins/aws/invokeLocal/index.js
@@ -21,14 +21,14 @@ const ServerlessError = require('../../../classes/Error').ServerlessError;
 const cachePath = path.join(cachedir('serverless'), 'invokeLocal');
 
 const envVarResolvers = {
-  'AWS::S3::Bucket': require('./resolvers/s3'),
-  'AWS::DynamoDB::Table': require('./resolvers/dynamodb'),
-  'AWS::SNS::Topic': require('./resolvers/sns'),
-  'AWS::SQS::Queue': require('./resolvers/sqs'),
-  'AWS::Lambda::Function': require('./resolvers/lambda'),
-  'AWS::SecretsManager::Secret': require('./resolvers/secrets'),
-  'AWS::SSM::Parameter': require('./resolvers/ssm-parameter'),
-  'AWS::StepFunctions::StateMachine': require('./resolvers/state-machine'),
+  'AWS::S3::Bucket': require('./resolvers/s3Bucket'),
+  'AWS::DynamoDB::Table': require('./resolvers/dynamoDbTable'),
+  'AWS::SNS::Topic': require('./resolvers/snsTopic'),
+  'AWS::SQS::Queue': require('./resolvers/sqsQueue'),
+  'AWS::Lambda::Function': require('./resolvers/lambdaFunction'),
+  'AWS::SecretsManager::Secret': require('./resolvers/secretsManagerSecret'),
+  'AWS::SSM::Parameter': require('./resolvers/ssmParameter'),
+  'AWS::StepFunctions::StateMachine': require('./resolvers/stepFunctionsStateMachine'),
 };
 
 let serviceConfig = null;
