+++ lib/plugins/aws/invokeLocal/resolvers/dynamoDbTable.js
