+++ lib/plugins/aws/invokeLocal/resolvers/lambdaFunction.js
