+++ lib/plugins/aws/invokeLocal/resolvers/s3Bucket.js
