+++ lib/plugins/aws/invokeLocal/resolvers/snsTopic.js
