+++ lib/plugins/aws/invokeLocal/resolvers/sqsQueue.js
