+++ lib/plugins/aws/invokeLocal/resolvers/ssmParameter.js
