+++ lib/plugins/aws/invokeLocal/resolvers/stepFunctionsStateMachine.js
